+++ src/StoreApi/Exceptions/NotPurchasableException.php
@@ -1,5 +1,5 @@
 <?php
-namespace Automattic\WooCommerce\Blocks\StoreApi\Utilities;
+namespace Automattic\WooCommerce\Blocks\StoreApi\Exceptions;
 
 /**
  * NotPurchasableException class.
@@ -7,6 +7,4 @@ namespace Automattic\WooCommerce\Blocks\StoreApi\Utilities;
  * @internal This API is used internally by Blocks, this exception is thrown when an item in the cart is not able to be
  * purchased.
  */
-class NotPurchasableException extends StockAvailabilityException {
-
-}
+class NotPurchasableException extends StockAvailabilityException {}
