+++ src/StoreApi/Exceptions/OutOfStockException.php
@@ -1,5 +1,5 @@
 <?php
-namespace Automattic\WooCommerce\Blocks\StoreApi\Utilities;
+namespace Automattic\WooCommerce\Blocks\StoreApi\Exceptions;
 
 /**
  * OutOfStockException class.
@@ -7,6 +7,4 @@ namespace Automattic\WooCommerce\Blocks\StoreApi\Utilities;
  * @internal This API is used internally by Blocks, this exception is thrown when an item in a draft order is out
  * of stock completely.
  */
-class OutOfStockException extends StockAvailabilityException {
-
-}
+class OutOfStockException extends StockAvailabilityException {}
